Consumindo serviço REST, enviando o json de pagamento, implementado o cliente REST.

+++ controllers/pagamentos.js
@@ -49,10 +49,10 @@ module.exports = function(app){
 
     app.post('/pagamentos/pagamento', function (req, res) {
 
-        req.assert("forma_de_pagamento",
+        req.assert("pagamento.forma_de_pagamento",
             "Forma de pagamento obrigatório").notEmpty();
 
-        req.assert("valor","Valor é obrigatório e dee ser um decimal")
+        req.assert("pagamento.valor","Valor é obrigatório e dee ser um decimal")
             .notEmpty().isFloat();
 
         var erros = req.validationErrors();
@@ -62,7 +62,7 @@ module.exports = function(app){
             return
         }
 
-        var pagamento = req.body;
+        var pagamento = req.body["pagamento"];
         console.log('Processando req de novo pagamento')
 
         pagamento.status = 'CRIADO'
@@ -78,25 +78,40 @@ module.exports = function(app){
             }else{
                 pagamento.id = resultado.insertId
                 console.log('pagamento criado')
-                res.location('/pagamentos/pagamento/' + pagamento.id)
 
-                var response= {
-                    dados_do_pagamento: pagamento,
-                    links:[
-                        {
-                            href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                            rel:"confirmar",
-                            method:"PUT"
-                        },
-                        {
-                            href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                            rel:"cancelar",
-                            method:"DELETE"
-                        },
-                    ]
+                if(pagamento.forma_de_pagamento == 'cartao') {
+                    var cartao = req.body["cartao"];
+                    console.log(cartao);
+
+                    var clienteCartoes = new app.servicos.clienteCartoes();
+                    clienteCartoes.autoriza(cartao, function(exception, request, response, retorno){
+                        console.log(201).json(retorno);
+                        return;
+                    });
+
+                    res.status(201).json(cartao);
+                    return;
+                } else {
+                    res.location('/pagamentos/pagamento/' + pagamento.id)
+
+                    var response= {
+                        dados_do_pagamento: pagamento,
+                        links:[
+                            {
+                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                rel:"confirmar",
+                                method:"PUT"
+                            },
+                            {
+                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                rel:"cancelar",
+                                method:"DELETE"
+                            },
+                        ]
+                    }
+
+                    res.status(201).json(response)
                 }
-
-                res.status(201).json(response)
             }
         })
     })
